Program/requirements.txt
ao3-api==2.3.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
DateTime==4.8
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
numpy==1.24.0
packaging==22.0
pathlib==1.0.1
Pillow==9.3.0
pyparsing==3.0.9
python-crontab==2.7.0
python-dateutil==2.8.2
pytz==2022.7
regex==2022.10.31
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.9.0
urllib3==1.26.13
warn==0.1.0
zope.interface==5.5.2
